requirements.txt
flask==2.3.2
opencv-python==4.8.1.78
imagehash==4.3.1
Pillow==9.5.0
numpy==1.25.2
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2

# Optional (useful for production deployments)
# gunicorn==20.1.0

